Fix handling of transition to SPDX headers
The old header extended two lines after the token by which it was
recognized, which never mattered before (because changes to the
copyright header would still leave it with those two lines at the end)
but, of course, they matter now. The precise matching of the end also
matters. The conversion to SPDX also added some start-comment markers
between the SPDX header and further commentary, where the old header
continued on to such, notably in the glgen-generated header files.

Task-number: QTQAINFRA-5013
Change-Id: I7b8c3e174a350a3509837f155001b79930f485ca
Reviewed-by: Lucie Gerard <lucie.gerard@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>