Fix harmonizing of old and new forms with common plausible precursor
Because skipping Q_DECL_NOTHROW and skipping noexcept are both valid,
a line which changed from the former to the latter was shown as having
had the former removed. In general, if any of the plausible precurors
of the prior and replacement lines coincide, we can just treat the
prior line as if it had not changed.

Fixes: QTBUG-104417
Change-Id: I5ef8503766575fe0be7361876fa89442bbdd631d
Reviewed-by: Ievgenii Meshcheriakov <ievgenii.meshcheriakov@qt.io>