Add a script to warn about missing pick-to branch targets
After branching Qt, it is necessary to ensure that all outstanding
changes which should target the new branch do so. This script
performs an on-demand sanity-check of the Pick-to commit message
footer on open changes. The script checks all major branches
for outstanding changes, so no configuration following branching
is necessary. If a change targets major.current-1, but not
major.current branch, a -1 sanity warning with relevant comment
will be posted.

See also sanity bot update:
125f1520ee41b018888e2e2652a0b1c299f93550

Task-number: QTQAINFRA-5036
Change-Id: I957076ee207d05bbf9324e6f2f637ad93747d51f
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>