fuzzing: Refactor build.sh to allow the use of local Qt checkout
The way the oss-fuzz Dockerfile is currently defined does not allow for
the use of a local checkout of Qt when building the fuzzers or the test
cases.

In order to fix that, the first step is to update this script so it is
ready for the new version of the Dockerfile but also keep it compatible
with the current version so that there's no break on the fuzzing side
until the new Dockerfile is merged by the oss-fuzz project.

Related: https://github.com/google/oss-fuzz/issues/7634

Change-Id: I98bfa4a3d83d1a3259eb4b7a33e839ededc54751
Reviewed-by: Robert Löhning <robert.loehning@qt.io>