Improve attention set handling for new cherry-picks
A user may only be added to a change's attention set if the user is
an owner, uploader, reviewer, or cc. As the cherry-pick bot
owns changes it creates, the original owner must be added to the
cherry-pick's reviewer set before attempting to flag them for
attention. This action is only performed when a pick has merge
conflicts. If the original owner does not have access to tqtc/ repos
but the cherry-pick was done to a private repo/branch,
a fallback process is necessary to locate a suitable user with
appropriate access.

Fixes: QTQAINFRA-5157
Fixes: QTQAINFRA-5173
Change-Id: Icb33915cc2e8594ef10fd351f37759323995eefe
Reviewed-by: Jukka Jokiniva <jukka.jokiniva@qt.io>