Waterfall cherry-picks from newer to older branches
The current behavior of blasting all requested picks at once is
causing some issues with apparent regressions in the event that
a change is cherry-picked to both an older branch and a newer one,
but the newer pick fails to merge. When a user upgrades versions,
it then appears that there has been a regression.

In order to reduce these types of incidents, the pick-to
targets are checked for any gaps before determing the pick order.
Any missing targets will be added to the Pick-to footer resulting from
the first cherry-pick. The latest feature branch specified in or added
to the footer will be picked first, and any remaining feature
or release branches are tacked on to a new Pick-to:
footer of the cherry-pick.

Change-Id: I8176395fcd082e235280fc0f7da06cf95190bfcd
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>