Add a parser for CMakeLists.txt files, to extract public headers
Give the name of the module on the command line, supply the file's
contents on stdin (so we can git show an old version without checking
it out) and it'll reply on stdout with the headers it found. Soon to
replace the sync.profile-based sources of API-defining headers.

Task-number: QTQAINFRA-4763
Change-Id: Ic6d4aa2b4c0160a0e995a35765fecc6e506b0403
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>