Cope with modules coming and going or being moved around
This may mean we have to consult more than one CMakeLists.txt file for
a single module. Since this leads to consulting both old and new
CMakeLists.txt files, the new header list may now contain some headers
from the old list, that we'll fail to check out; don't die on that,
just suppress the same warning as for the old header checkout to the
new tag.

Change-Id: Ief68888ead67f09d54abd6ef9c460c4701c099bb
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>