Add a plugin to choose fix versions and close Jira issues
This plugin to the cherry-pick bot monitors for change-merged
events and for changes which include the "Fixes: " or "Task-Number: "
keyword. The plugin will tag the issue with the merged SHA in both
cases.

If the Fixes: keyword is used, the plugin will then attempt to
determine a suitable fix version. It will also close the jira ticket
if ticket has not been reopened by a user after the bot has
previously closed it once.

Fixes: QTQAINFRA-5145
Fixes: QTQAINFRA-5478
Change-Id: Icdba6dc421d4353b459539b5ff68274229979b87
Reviewed-by: Daniel Smith <Daniel.Smith@qt.io>