Mark the module as free of Q_FOREACH, except where it isn't
The density of Q_FOREACH uses is not very high here, but too high for
this author to tackle in a short amount of time. Since they're
concentrated in just a few TUs, pick a different strategy:

Mark the whole module with QT_NO_FOREACH, to prevent new uses from
creeping in, and whitelist the affected TUs by #undef'ing
QT_NO_FOREACH locally, at the top of each file. For TUs that are part
of a larger executable, this requires these files to be compiled
separately, so add them to NO_PCH_SOURCES (which implies
NO_UNITY_BUILD_SOURCES, too). Created QTBUG-115806 to keep track of
this.

Needed to port the use in a header, as otherwise we'd have to put all
TUs into NO_PCH_SOURCES.

Task-number: QTBUG-115806
Change-Id: I62bdcabb0f34844b9b9935b431a31ef2b3f1c09c
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>