Exclude QDoc's expected output from license check
Depending on the configuration, QDoc may generate XML files as part of
its output. These XML files do not pass the license check, as they
(correctly) do not contain an SPDX license header.

Exclude QDoc's expected output directory from the license check, as
the contents of this directory is expected to change frequently, and as
these files cannot be subject to licensing.

Task-number: QTBUG-114937
Change-Id: Idb98926f0c38bac7a660c72895cf853ea8ff2957
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>