Exclude QDoc "QualTypeNames.h" from the license test
"QualTypeNames.h" is a recently introduced header that has been imported
from the LLVM project and should not incur into the license checking.

It is currently embedded directly into QDoc source code acting as a drop
in replacement for the one that is exported from upstream, so that it is
not under a "3rdparty" directory.

Hence, a direct exception for the file is no included to avoid issues in
CI.

Task-number: QTBUG-111580
Change-Id: Ia2affcf8fc7300855e5fe94753f3c2e522bc5b9a
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 9b782b3dd60e4b18056ae5b769c0ce57a2af615d)
Reviewed-by: Toni Saario <toni.saario@qt.io>