Move vale config
Move the vale linter config so that it is
easily accessible to those who want to use
it. To use this config with the vale linter
v2.29 or later, either set the VALE_CONFIG_PATH
environment variable with the abosolute path to
the .vale.ini file, or manually pass the path
with every call to vale using the -c command-line
option.

Note: Run vale sync the first time you use
the config. It downloads latest version of the
Microsoft style guide rules.

Task-number: QTBUG-119296
Change-Id: I09ef3d85a73aaa8e18203d8092999d193fafd98f
Reviewed-by: Andreas Eliasson <andreas.eliasson@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>