Add new functionality to remove Code-Review +2 for dev changes
When a branching event occurs, this script now also removes Code-Review
+2 votes from changes on the dev branch which do not target the new
latest stable branch. This is to force a re-evaluation of the change
to determine if the change should be pushed through to the new FF stable
branch or if the change can simply merge normally to dev for the next
stable release.

This change also fixes a minor, but breaking bug in the original script
where the regex used to parse branches was incorrectly used.

Change-Id: Idadec12ca8d958f7d86e83bfeaaa123d121a7bab
Reviewed-by: Daniel Smith <daniel.smith@qt.io>