Work round dulwich changing its IndexEntry internals
The pseudo-constructor for replacement IndexEntry objects needed to
get the right number of parameters, for which it needed to know how
many that would be. Inconveniently, that meant accessing an internal
data field that was named _fields; but from dulwich 0.21 IndexEntry is
an @dataclass, so the relevant member is __dataclass_fields__ instead.
Remain backwards-compatible for now, as Jani has the older version
still.

Task-number: QTBUG-119952
Change-Id: Ia4b0bc336442324e228713da0c597dcc383b843a
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Ievgenii Meshcheriakov <ievgenii.meshcheriakov@qt.io>