Fix resetboring's failure to spot hspace-only changes
The 6.7 review for QSql included a file where all that had changed was
some s/Type& name/Type &name/ fixes. This arose because there was no
other change to the lines in question and the initial scan of the new
lines thus concluded nothing boring had changed, so there was no point
recording the line for comparison to old lines. Retain at least one
representation of each non-trivial line to match, so that we do catch
cases where the only boring change is to spacing. Package the
non-trivial line check (previously used when reversing strip) in a
function since we're reusing it; and include backslash in its list of
trivia, since that arises in trivial lines of a macro body.

Change-Id: I10443bd1494932696f176f90b663035a772ca805
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>