Prefer somewhat similar old lines to match a new one
When resetboring has more than one (non-trivial) line in the old code
that could be a match to a new line, prefer an exact match if present
or one with the same indent otherwise. It's still crude, but slightly
less so than just taking the first - and marginally reduces noise.

Change-Id: I913fb51363cf81efea37ac0a24c698bfa3572200
Reviewed-by: Daniel Smith <daniel.smith@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>