Fix cache population for branch data in jira_closer
The jira closer bot performs a check on branching date
to determine if a given change merged to a branch before
a given branching date. This is done with MomentJS, and a
check which validates that cached datetime is a Moment object
fails due to improperly populating the local cache from database
upon startup. This ultimately leads to the wrong version being chosen
when a change is merged to a branch after the branching date.

Fixes: QTQAINFRA-6064
Change-Id: I1ef32f3cae9ee5bc38c2b5b7f5b3a77adc4f668e
Reviewed-by: Daniel Smith <daniel.smith@qt.io>