Enforce SPDX in-source expressions to match Qt Project rules
CI will now check that in-source SPDX-License-Expressions
do match the rules by the Qt Project as outlined in QUIP-18:

https://contribute.qt-project.org/quips/18

The exact rules are looked up in a file licenseRule.json at
the base directory of the tested Qt module. Such
licenseRule.json files have been added to all qt5 repositories
from 6.7 onwards.

If the file licenseRule.json does not exist, a warning
is shown, but the test will not fail. This allows older
branches to still pass the CI.

If the test fails due to a new file, you can either fix the
SPDX-License-Expression in the file, or - if warranted -
add an exception to the repository's licenseRule.json.
The file format for licenseRule.json is explained in the
sources.

Task-number: QTBUG-121039
Change-Id: Ia08725ea492c196355b45d83b185f1d6c7dea781
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>