Modify pick behavior when the expected parent has not been picked yet
Current behavior of picking to the first-found parent in a
relation chain can sometimes cause issues with broken dependencies.
This change modifies the behavior to always wait for the expected
parent to be picked before picking the child. If the wait times out,
the bot will fall back to the old behavior of walking the chain
to find the nearest suitable parent.

Fixes: QTQAINFRA-6167
Change-Id: I3074988b10e4efad7307770ef5ac218bb39d5066
Reviewed-by: Daniel Smith <daniel.smith@qt.io>