Extend cherry-pick bot to support using submit in select repos
Some repos are in submit-only mode and do not use the CI.
This patch adds support for using the submit command instead,
based on configuration in the config.json or environment variables.
This config is only made available per-repo, and cannot be used
to make a specific branch submit-only.

Using configuration over deductive or fallback logic is intended to
be more explicit to reduce security concerns of unreviewed code
being hard-submitted in unapproved repos.

Task-number: QTQAINFRA-6157
Change-Id: If2d0e538369c470e0753d311e183ea736981b63c
Reviewed-by: Daniel Smith <daniel.smith@qt.io>