Fix detection of tqtc/lts- and lts- prefixed branches in jira closer
Due to how the jira closer matches fix version based on raw
version strings, it was incorrectly checking gerrit for the
existence of tqtc/lts- and lts- prefixed branches as the raw
branch without the prefix. This caused the bot to believe that
the target branch was not yet created, and assign the wrong fix
version to the jira issue.

Fixes: QTQAINFRA-6064
Change-Id: If3879ba0a696eca613432ee01d633bb347accc99
Reviewed-by: Daniel Smith <daniel.smith@qt.io>