Add API endpoint for calculating JIRA FixVersion for a gerrit changes
This patch introduces /jiracloser/fixversion endpoint which
calculates the JIRA FixVersion for a gerrit change. The
The endpoint accepts only the unique gerrit change number in the
"change" query parameter. Gerrit Change Ids are not supported.
The endpoint responds with 200 status code and a JSON body of
one or more JIRA fixVersions.

Change-Id: Iddaf3b6bfb4527c1d384d77b5ee8ff8794fd0cb3
Reviewed-by: Daniel Smith <daniel.smith@qt.io>
Reviewed-by: Audun Sutterud <audun.sutterud@qt.io>