Skip closed branches in the cherry-picking automation
This patch changes the behavior of the cherry-pick bot to skip branches
that are already closed for new changes, instead of halting upon
encountering such. This behavoir change comes as a result of the move
to the waterfall model, where the pick bot is expected to pick targets
in a sequential manner rather than simultaneously. This works around
the possible scenario where a change on dev is merged with valid pick
targets, but during the waterfall process, one of those targets is
closed. Now, the bot will still post a warning message about the closed
branch, but then continue to pick to the next target regardless.

Fixes: QTQAINFRA-6287
Fixes: QTQAINFRA-6288
Change-Id: I91d9e066b4bcba1763049272fdf26c3b267d2688
Reviewed-by: Daniel Smith <daniel.smith@qt.io>