Improve API for Jira Closer bot to utilize fallbacks
API specific: If an issue is queried and either does not have
a Fixes: footer, or querying that issue results in no fix versions,
the bot will now attempt to find the fix version from a a default
QTBUG query. This only applies if the change is in the qt/ namespace.

This does not functionally alter which fix version should be applied
to the change, but allows for guessing fix versions via the API for
changes which do not specify a relationship to a Jira issue.

Fixes: QTQAINFRA-6312
Change-Id: I6237ca4fa79416afad515f39d8bb104290a6391f
Reviewed-by: Daniel Smith <daniel.smith@qt.io>