CMake: Disable generating SBOM for qtqa
The repo doesn't install anything by default, but because we end up
installing the sbom file, we start trying to build the tests in the CI,
and that fails due to unrelated issues that need to be fixed
separately.

To avoid the issue, explicitly disable generating SBOMs for this repo,
even if the feature is enabled in the platform config.

Task-number: QTBUG-122899
Change-Id: I582e17db699c8b266b4f351ff3cb1a003d8db481
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>