Android: Add a script to generate Java API review
The script uses javadoc to generate java files that contain only
the API description, and its documentation. The script looks for
various Java sources under the qt5 super repo path and generates API
description for specific package names provided here, these pre-defined
Qt for Android package names. This process is done for the previous Qt
release and it's committed as a base to-be-ignored commit. The same is
done for Qt release being reviewed and a commit is created on top.

The script can push changes to Gerrit, and automatically abandon the
base commits. Also, it can update existing API review patches, given
that the local git branch that was used to push before still exists with
valid commits.

Optionally a Jira task number can be provided.

Fixes: QTBUG-126352
Change-Id: Ib5cc7c560e320dbda9e9df5104c06161dccd7561
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>