Add VM network test
Virtual machines in COIN have occassional network failures for unknown
reasons. This infrastructure flakiness makes tests fail, which depend
on QtNetwork. It triggers research, that consumes R&D resources, to
exclude regressions in QtNetwork.

Testing for potential regressions is inefficient in a potentially
unstable virtual machine.

Implement a network test application, that returns success if a VM has
stable network. Use tst_QDnsLookup::lookup() as a template, because it
- relies on TCP being available
- verifies internal lookups, that exist only in CI environment.

Add command line options, to cover diagnose and production needs.

The application has to be provisioned after having been built against
a stable/released branch of Qt - not the branch that is subject to CI
testing.

Document building and running requirements in a README.

Task-number: QTQAINFRA-6796
Change-Id: I541a7113b6ec223041442f590964fe15d562ef9c
Reviewed-by: Timur Pocheptsov <timur.pocheptsov@qt.io>