Add checking source SBOM licensing
tst_licenses.pl can now be used to check the license associated
to each file present in the source SBOM. For this check to trigger
the path to the source SBOM has to be passed via -sbom argument.

A direct check of the licensing documented in source SBOM is necessary
because tst_licenses.pl does not check the license usage in file
for all files. Also, a file could have correct licensing documented in
file, and thus pass the license test, but have this licensing wrongly
overwritten in a REUSE.toml (the file read to create the source SBOM).

Since the logic between the licenses is lost in the source SBOM,
the check ignores the logic.
For the moment, files within 3rdparty folders are excluded from the
check.

Pick-to: 6.9
Task-number: QTBUG-131434
Change-Id: Icd924f2fd0704aed9686c822c433e99ca6ecd309
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
(cherry picked from commit ab757eda7cf90681419d94a3990f8e212df38bfd)
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>