Android: Move the Java API signature generator under qtqa repo
Previously, the javadoc doclet used to generate the API signature and
description of the Java APIs was using a 3rd party project under a
separate repo, that's not ideal. Now, the doclet is moved under qtqa.

Prior to this, it was expected that the binary of that doclet is
provided to the script manually. Now, it would be compiled and used
directly internally by the script. For that reason, we no longer need
the --doclet argument.

The expected output now is a single file describing the API of a given
module instead of a directory with multiple files representing various
classes in the API.

Additionally, the doclet was using old JDK APIs that are deprecated. It
used to require JDK 11, with this change, newer JDKs can be used because
we're using newer APIs. For that reason we no longer need the --javadoc
argument.

Change-Id: Ifc1568a7a716d3eb740dba1ea3b5a114f2201ebe
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit f9555f998a4d4f80b51c0dcb77ce28c9ed2d9031)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 93c52061bff574ea383e8638443f6e0615c74c0f)
(cherry picked from commit cedd77ec89370d6576be87003500cdc72ae8f8db)